refactor: combine exports into barrel modules
