+++ src/App.tsx
@@ -1,9 +1,6 @@
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
 
-import Home from './pages/Home';
-import NotFound from './pages/NotFound';
-import About from './pages/About';
-import FormPage from './pages/FormPage';
+import { Home, NotFound, About, FormPage } from './pages';
 
 export default function App() {
   return (
