+++ src/assets/index.ts
@@ -1,0 +1,6 @@
+export { default as games } from './data/games.json';
+export { default as genres } from './data/genres.json';
+export { default as favorite_empty } from './img/favorite_empty.svg';
+export { default as favorite_filled } from './img/favorite_filled.svg';
+export { default as metacritic_icon } from './img/metacritic_icon.svg';
+export { default as search_icon } from './img/search.svg';
