+++ src/components/header/Header.tsx
@@ -1,4 +1,4 @@
-import Navbar from './Navbar';
+import { Navbar } from '../';
 
 type HeaderProps = {
   currentPage: string;
