+++ src/components/index.ts
@@ -1,0 +1,8 @@
+export { default as Card } from './card/Card';
+export { default as CardButtons } from './card/CardButtons';
+export { default as CardContainer } from './card/CardContainer';
+export { default as Navbar } from './navbar/Navbar';
+export { default as Form } from './form/Form';
+export { default as Select } from './form/Select';
+export { default as Header } from './header/Header';
+export { default as SearchBar } from './searchBar/SearchBar';
