+++ src/pages/NotFound.tsx
@@ -1,4 +1,4 @@
-import Header from '../components/Header';
+import { Header } from '../components';
 
 export default function NotFound() {
   return (
