+++ src/pages/index.ts
@@ -1,0 +1,4 @@
+export { default as About } from './About';
+export { default as FormPage } from './FormPage';
+export { default as Home } from './Home';
+export { default as NotFound } from './NotFound';
